DB, S3 테스트 완료 및 폴더구조 수정 #5
