+++ node-restful/app.js
@@ -1,3 +1,5 @@
+require('dotenv').config()
+
 const express = require('express')
 const app = express()
 const port = process.env.SERVER_PORT
